blessing finishes ticket 2

+++ models/index.js
@@ -6,7 +6,7 @@ exports.selectCategories = () => {
 }
 exports.selectReviews = () => {
     const query = `
-    SELECT reviews.*, COUNT(comments.review_id)  AS comment_count
+    SELECT reviews.*, COUNT(comments.review_id)::INT AS comment_count
     FROM reviews 
     LEFT JOIN comments ON comments.review_id = reviews.review_id
     GROUP BY reviews.review_id
